refactor : product 패키지 이동
- domain -> domain/product
