+++ src/test/java/cart/domain/product/NameTest.java
@@ -1,19 +1,20 @@
-package cart.domain;
+package cart.domain.product;
 
+import cart.domain.product.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class ImageTest {
+class NameTest {
 
     @ParameterizedTest
     @ValueSource(strings = {"", " ", "  "})
-    @DisplayName("validateIsBlank() : url을 공백으로 입력 시 IllegalArgumentException가 발생합니다.")
+    @DisplayName("validateIsBlank() : 이름을 공백으로 입력 시 IllegalArgumentException가 발생합니다.")
     void test_validateIsBlank_IllegalArgumentException(final String value) throws Exception {
         //when & then
-        assertThatThrownBy(() -> new Image(value))
+        assertThatThrownBy(() -> new Name(value))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
