+++ src/test/java/cart/domain/product/PriceTest.java
@@ -1,5 +1,6 @@
-package cart.domain;
+package cart.domain.product;
 
+import cart.domain.product.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
